requirements.txt
Flask==2.3.2
Flask_Cors==3.0.10
Pillow==11.0.0
